Constrain the width of the grid.
Summary: We center the grid using javascript based on the available width after we have filled it with columns. Due the the fact that grid items are absolutely positioned we render them within a display: relative div. The display: relative div is then translated left based on the amount of pixels needed for centering. We also need to set this width on a non-relative ancestor in order to avoid a horizontal scrollbar.