Grid integration testing (#19)
This adds a new type of integration test to gestalt. Initially this adds a single test which verifies that grid items are positioned correctly after re-hydrating from a server-side render. There's quite a bit of boilerplate required to get the server side rendering working, but currently that's encapsulated in a folder which means we could probably remove it easily down the road.