BoxGrid (#182)
This is a grid which uses a simple algorithm to find the lowest available space across multiple columns. The packer keeps track of open spaces, and splits an open space when we insert something into the grid. It could compact more space if we perform item sorting, but for now we're just inserting into the first available slot.