[BoxGrid] Packing algorithm fixes (#189)
Previously the findNextShortest function would return the lowest column it could find after a given search height. The position method would then increment the search height to continue searching. This meant that we could miss checking for columns that were the same height as we would skip over them. This commit makes findNextShortest return an array of slots for a given height, instead of a single slot.