Remove height calculation resetting. (#226)
The containerHeight state specifies how tall the grid is. This allows consumers of the grid to position items under it. It is not necessary to set this to zero as it will be re-calculated after it fully renders. This should also save the browser from doing unnecessary work, and in some cases overlapping two reflowing grids.