Expose a reflowGrid method on GroupAvatar. (#285)
In order to avoid measuring the DOM for every render() call, the grid maintains a cache of items. This cache will unfortunately break subtree diffing and currently needs to be invalidated manually. This patch adds a reflowGrid method on GroupAvatar which consumers can use to clear the cache when updating GroupAvatar props.